requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
weaviate-client
sentence-transformers
textblob